backend/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
numpy==1.24.3  # NumPy 2.0 이전 버전 지정
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.3
faiss-cpu==1.7.4
sentence-transformers==2.2.2
httpx==0.25.2
pydantic==2.5.2
python-dotenv==1.0.0
huggingface-hub==0.19.4
langgraph>=0.0.20